% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_cleaner.R
\name{clean_html_text}
\alias{clean_html_text}
\title{Очистка HTML текста}
\usage{
clean_html_text(html_string)
}
\arguments{
\item{html_string}{Сырой HTML код (строка)}
}
\value{
Чистый текст
}
\description{
Удаляет теги, скрипты, стили и лишние пробелы из текста новости.
}
