% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_storage.R
\name{init_db_schema}
\alias{init_db_schema}
\title{Инициализация таблицы}
\usage{
init_db_schema(con)
}
\arguments{
\item{con}{Соединение с БД}
}
\description{
Создает таблицу news_analysis, если она не существует.
}
