% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dashboard.R
\name{run_dashboard}
\alias{run_dashboard}
\title{Запуск аналитического дашборда}
\usage{
run_dashboard()
}
\description{
Ищет файл dashboard.qmd внутри пакета и запускает его через Quarto.
Автоматически подгружает .env из текущей рабочей директории.
}
