% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_storage.R
\name{save_article_to_db}
\alias{save_article_to_db}
\title{Сохранение результатов}
\usage{
save_article_to_db(con, news_row, ai_result, clean_text)
}
\arguments{
\item{con}{Соединение с БД}

\item{news_row}{Строка с данными новости (из TT-RSS)}

\item{ai_result}{Список с результатами от YandexGPT (category, summary, score)}

\item{clean_text}{Очищенный текст новости (принимаем, даже если не пишем в эту таблицу)}
}
\description{
Сохраняет новость и результат работы ИИ в базу.
}
