% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttrss_client.R
\name{ttrss_get_unread}
\alias{ttrss_get_unread}
\title{Получение непрочитанных новостей}
\usage{
ttrss_get_unread(base_url, sid, feed_id = -4, limit = 5)
}
\arguments{
\item{base_url}{Адрес API}

\item{sid}{Session ID}

\item{feed_id}{ID ленты (-4 = все ленты)}

\item{limit}{Количество статей (по умолчанию 5)}
}
\value{
Data frame со статьями или NULL
}
\description{
Скачивает заголовки и контент свежих статей.
}
