% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttrss_client.R
\name{ttrss_login}
\alias{ttrss_login}
\title{Авторизация в TT-RSS}
\usage{
ttrss_login(base_url, user, password)
}
\arguments{
\item{base_url}{Адрес API (например, http://ip/api/)}

\item{user}{Логин}

\item{password}{Пароль}
}
\value{
Строка с session_id
}
\description{
Получает session_id для дальнейшей работы с API.
}
