% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttrss_client.R
\name{ttrss_mark_read}
\alias{ttrss_mark_read}
\title{Пометить новость как прочитанную}
\usage{
ttrss_mark_read(base_url, sid, article_id)
}
\arguments{
\item{base_url}{Адрес API}

\item{sid}{Session ID}

\item{article_id}{ID новости (число или вектор чисел)}
}
\value{
TRUE если успешно
}
\description{
Отправляет запрос updateArticle, чтобы убрать статус Unread.
}
